Ask Optiq: review pass - isolate the feature, fix found bugs, tighten the script contract
Isolation. The assistant and scripting code no longer sits inside shared core
files. The Settings dialog's Assistant page and its credential-store access,
the TraceView methods that exist only for OptiqActions, and the DataProvider
script path all moved into conditionally compiled files, so a feature-off
build does not carry them. Reading and writing the endpoint list deliberately
stays in rocprofvis_settings_manager.cpp: it runs unguarded, so moving it
would erase a saved configuration in a build with the assistant off.

Bugs found and fixed:
- The Python runtime reported success from a later Init() when startup had
  already failed, parking every subsequent script on a future that could
  never complete.
- JSON numbers from the model were cast straight to uint64_t, which is
  undefined for NaN, infinity, or anything past the type's range.
- ScriptEditor ignored a failed CancelScript() and sat on Running forever.
- ComputeView installed a this-capturing progress callback with an empty
  destructor; TraceView cleared six callbacks but not that one.
- SelectAnalysisTab wrote the global show_details_panel but applied it to one
  trace, desyncing every other open tab.
- PyThreadState_SetAsyncExc ignored a >1 return, which CPython requires be
  reverted.
- annotate mixed a model-supplied timestamp with a live-read view range and
  reported unconditional success, so a stale anchor was described as the
  selected interval. It now validates against the trace and says where the
  note actually landed.

Duplication and dead code. View::to_lower_copy, TrimCopy and ToLowerAscii all
duplicated existing Core::String helpers. Removed the PanelBacking enum whose
only consumer asked one question of it, an unreachable warmup branch, an
unused result code, and a completion event posted to a listener that tab-close
had already destroyed.

Script contract. The model's only API reference is the tool schema, while the
real documentation lives in PYTHON.md, which nothing loads - so it guessed
argument types and got them wrong. The schema now states that tracks= takes
Track objects rather than ids, that where/group are strings, that row cells
are often str, that categories are trace-defined, and that the same GPU work
appears on several tracks so naive concatenation double-counts. The matching
binding errors say how to correct them.

Also: the Script tab's source/result split is now draggable, using the same
splitter as the Ask Optiq dock.
