+++ .agents/UI.md
@@ -419,7 +419,6 @@ AppWindow (singleton, RocWidget)
 +-- SshTestDialog            : optional dev-only remote trace UI
 +-- AppMonitor               : singleton; polls background controller operations
 +-- LogViewer                : singleton production log overlay
-+-- ScriptEditor             : singleton Python analysis overlay (scripting flag)
 +-- NotificationManager     : singleton, drawn last (toasts overlay)
 +-- DebugWindow             : singleton, dev mode only
 ```
@@ -784,18 +783,26 @@ subclassing this; don't roll your own.**
 
 ### 7.15 `rocprofvis_script_editor.{h,cpp}` - Python analysis editor
 
-Gated on `ROCPROFVIS_ENABLE_SCRIPTING`. `class ScriptEditor` is a
-lazy singleton like `LogViewer`: floating `ImGui::Begin` overlay,
-not a docked column. `AppWindow` destroys it, renders it after
-`LogViewer`, and binds `View > Show Script Editor` to `VisiblePtr()`.
-`TraceView` and `ComputeView` toolbars call `RenderToolbarButton()`
-("Script") next to Ask Optiq.
+Gated on `ROCPROFVIS_ENABLE_SCRIPTING`. `class ScriptEditor` is the
+**Script tab** of the details panel, built and owned by `AnalysisView`
+beside Event Table / Top Events / Annotations. It is a plain
+`RocWidget` - no `ImGui::Begin`, no singleton, no visibility flag - so
+there is one editor per trace and `Run` reaches its own
+`DataProvider` and `TimelineSelection` without asking which tab is in
+front. Compute traces have no `AnalysisView`, so they have no Script
+tab.
 
 Run sends the source string through `DataProvider::ExecuteScript`
-(current tab, selected tracks / time range). Cancel uses
-`CancelScript`. Load/Save go through `AppWindow` file dialogs with a
-`.py` filter. Output is `optiq.result.text` / the error string. No
-syntax highlighting; result tables are Phase 2.
+(selected tracks / time range, or all tracks and the full trace).
+Cancel uses `CancelScript`. Load/Save go through `AppWindow` file
+dialogs with a `.py` filter. Output is `optiq.result.text` / the error
+string. No syntax highlighting; result tables are Phase 2.
+
+Ask Optiq offers a script through `OptiqActions::ProposeScript`, which
+fills this tab and selects it; nothing runs until the user presses Run
+or Reject. `ScriptApproval` is that state machine, and
+`ScriptExecuteCompleteEvent` is filtered by trace source id because
+tab-close cleanup posts one too.
 
 ## 8. Track Item Hierarchy
 
@@ -2567,7 +2574,7 @@ adding **anything** new, check this list and reuse if at all possible.
 | Keep the lazy render loop awake               | `RootView::WantsContinuousRender` / `AppWindow::WantsContinuousRender`                         |
 | Monitor controller operations without blocking| `AppMonitor::AddOperation` + typed status events                                               |
 | Show application logs                         | `LogViewer` (not the developer-only `DebugWindow`)                                             |
-| Run an in-app Python analysis script          | `ScriptEditor` (`View > Show Script Editor` + toolbar; `ROCPROFVIS_ENABLE_SCRIPTING`)          |
+| Run an in-app Python analysis script          | `ScriptEditor`, the details panel's Script tab, owned by `AnalysisView` (`ROCPROFVIS_ENABLE_SCRIPTING`) |
 | Add a hotkey                                  | Add to `HotkeyActionId`, declare info, read with `WasActionTriggered / IsActionHeld`           |
 | Disambiguate clicks across timeline layers    | `TimelineFocusManager::RequestLayerFocus / EvaluateFocusedLayer`                               |
 | Encode a request ID                           | `RequestIdBuilder::MakeRequestId(...)` or `MakeTrackDataRequestId(...)`                        |
