+++ src/view/src/agenticprofiling/rocprofvis_ai_actions.cpp
@@ -15,6 +15,7 @@
 #include "rocprofvis_settings_manager.h"
 #include "rocprofvis_timeline_selection.h"
 #include "rocprofvis_trace_view.h"
+#include "rocprofvis_utils.h"
 #include "widgets/rocprofvis_log_viewer.h"
 #include "widgets/rocprofvis_notification_manager.h"
 #ifdef ROCPROFVIS_ENABLE_SCRIPTING
@@ -29,19 +30,10 @@ namespace View
 namespace
 {
 
-// True when a time range is one we can act on.
-//
-// The finiteness test is the point of this function. Every comparison against
-// NaN is false, so a bare "end_ns <= start_ns" guard lets NaN and infinity
-// straight through, and these numbers come from the model.
-bool
-IsUsableRange(double start_ns, double end_ns)
-{
-    return std::isfinite(start_ns) && std::isfinite(end_ns) && end_ns > start_ns;
-}
-
 // True when a single point in time is one we can act on. Same reasoning as
-// IsUsableRange, for the arguments that are an instant rather than a span.
+// is_usable_time_range, for the arguments that are an instant rather than a
+// span: every comparison against NaN is false, and these numbers come from the
+// model.
 bool
 IsUsableInstant(double time_ns)
 {
@@ -152,9 +144,18 @@ PanelFromKey(const std::string& key)
     return OptiqPanel::kUnknown;
 }
 
+// The minimap and annotations are drawn by a TraceView widget and persisted
+// nowhere, so they cannot be touched without the active trace. Every other
+// panel is backed by AppWindowSettings and follows the View menu, updating
+// every open system trace even when a compute tab is in front.
+bool
+NeedsActiveTraceView(OptiqPanel panel)
+{
+    return panel == OptiqPanel::kMinimap || panel == OptiqPanel::kAnnotations;
+}
+
 }  // namespace
 
-// Binds to one trace's view objects. Any of them may be null.
 OptiqActions::OptiqActions(DataProvider*      data_provider,
                            TimelineSelection* timeline_selection,
                            ComputeSelection*  compute_selection, TraceView* trace_view)
@@ -164,7 +165,6 @@ OptiqActions::OptiqActions(DataProvider*      data_provider,
 , m_trace_view(trace_view)
 {}
 
-// Resolves whatever the model called a panel to one of ours.
 OptiqPanel
 OptiqActions::PanelFromName(const std::string& name)
 {
@@ -179,7 +179,6 @@ OptiqActions::PanelFromName(const std::string& name)
     return PanelFromKey(StripPanelFiller(key));
 }
 
-// The canonical name of a panel, for reporting back what was changed.
 const char*
 OptiqActions::PanelName(OptiqPanel panel)
 {
@@ -197,7 +196,6 @@ OptiqActions::PanelName(OptiqPanel panel)
     }
 }
 
-// Every panel name, for the tool schema and for bad-argument replies.
 std::string
 OptiqActions::PanelNameList()
 {
@@ -209,61 +207,6 @@ OptiqActions::PanelNameList()
            "\"view\" or \"panel\"";
 }
 
-// Where a panel's visibility actually lives, which is what decides whether a
-// live TraceView is needed to change it.
-//
-//  - kTraceViewOnly       drawn by a TraceView widget, nothing persisted.
-//  - kSettingAndTraceView persisted in AppWindowSettings and mirrored on a
-//                         widget that was seeded from it at construction, so
-//                         both have to move: the setting is what survives a
-//                         restart, the widget is what redraws now.
-//  - kSettingAndAppWindow persisted and mirrored on the AppWindow layout.
-//  - kSettingOnly         persisted only; whoever draws it re-reads the flag
-//                         every frame, so there is no widget to miss.
-//  - kLogViewer           its own singleton, independent of any trace.
-//
-// Only trace-local panels need the active tab to be a TraceView. Setting-backed
-// panels follow the View menu: they update every open system trace even when a
-// compute tab is in front.
-namespace
-{
-
-enum class PanelBacking
-{
-    kTraceViewOnly,
-    kSettingAndTraceView,
-    kSettingAndAppWindow,
-    kSettingOnly,
-    kLogViewer
-};
-
-PanelBacking
-BackingOf(OptiqPanel panel)
-{
-    switch(panel)
-    {
-        case OptiqPanel::kMinimap:
-        case OptiqPanel::kAnnotations: return PanelBacking::kTraceViewOnly;
-        case OptiqPanel::kHistogram:
-        case OptiqPanel::kTopology:
-        case OptiqPanel::kDetails:     return PanelBacking::kSettingAndTraceView;
-        case OptiqPanel::kToolbar:     return PanelBacking::kSettingAndAppWindow;
-        case OptiqPanel::kSummary:     return PanelBacking::kSettingOnly;
-        case OptiqPanel::kLogViewer:   return PanelBacking::kLogViewer;
-        default:                       return PanelBacking::kSettingOnly;
-    }
-}
-
-// True when this panel cannot be touched without the active TraceView.
-bool
-NeedsActiveTraceView(OptiqPanel panel)
-{
-    return BackingOf(panel) == PanelBacking::kTraceViewOnly;
-}
-
-}  // namespace
-
-// Opens or closes a panel, the same as ticking its View-menu item.
 bool
 OptiqActions::ShowPanel(OptiqPanel panel, bool visible)
 {
@@ -395,21 +338,18 @@ OptiqActions::IsPanelVisible(OptiqPanel panel, bool& visible_out) const
     }
 }
 
-// True when the timeline actions are usable, i.e. a system trace is in front.
 bool
 OptiqActions::HasTimeline() const
 {
     return m_data_provider != nullptr && m_timeline_selection != nullptr;
 }
 
-// True when the compute actions are usable, i.e. a compute trace is in front.
 bool
 OptiqActions::HasCompute() const
 {
     return m_data_provider != nullptr && m_compute_selection != nullptr;
 }
 
-// The event source id the rest of the app filters on, which is the trace path.
 std::string
 OptiqActions::SourceId() const
 {
@@ -417,7 +357,6 @@ OptiqActions::SourceId() const
                                       : std::string();
 }
 
-// The open traces, by the tab label the user sees.
 std::vector<std::string>
 OptiqActions::ListTabs() const
 {
@@ -437,7 +376,6 @@ OptiqActions::ListTabs() const
     return names;
 }
 
-// The label of the trace currently in front.
 std::string
 OptiqActions::ActiveTab() const
 {
@@ -450,7 +388,6 @@ OptiqActions::ActiveTab() const
     return active != nullptr ? active->m_label : std::string();
 }
 
-// Brings a trace tab to the front by name.
 bool
 OptiqActions::SelectTab(const std::string& name)
 {
@@ -473,7 +410,6 @@ OptiqActions::SelectTab(const std::string& name)
     return true;
 }
 
-// The details panel's inner tabs, by label.
 std::vector<std::string>
 OptiqActions::ListAnalysisTabs() const
 {
@@ -481,14 +417,12 @@ OptiqActions::ListAnalysisTabs() const
                                    : std::vector<std::string>();
 }
 
-// The label of the details tab currently showing.
 std::string
 OptiqActions::ActiveAnalysisTab() const
 {
     return m_trace_view != nullptr ? m_trace_view->ActiveAnalysisTab() : std::string();
 }
 
-// Selects a details tab, opening the panel if it was hidden.
 bool
 OptiqActions::SelectAnalysisTab(const std::string& name)
 {
@@ -527,7 +461,6 @@ OptiqActions::ClearScriptProposal()
 }
 #endif
 
-// Shows or hides the arrows linking an event to what it launched or waited on.
 bool
 OptiqActions::SetFlowArrowsVisible(bool visible)
 {
@@ -539,7 +472,6 @@ OptiqActions::SetFlowArrowsVisible(bool visible)
     return true;
 }
 
-// Reads whether the flow arrows are currently drawn.
 bool
 OptiqActions::AreFlowArrowsVisible(bool& visible_out) const
 {
@@ -551,7 +483,6 @@ OptiqActions::AreFlowArrowsVisible(bool& visible_out) const
     return true;
 }
 
-// Switches the flow arrows between fanning out and chaining through the sequence.
 bool
 OptiqActions::SetFlowRenderChained(bool chained)
 {
@@ -563,7 +494,6 @@ OptiqActions::SetFlowRenderChained(bool chained)
     return true;
 }
 
-// Zooms back out to the whole trace, the same as the Reset View button.
 bool
 OptiqActions::ResetView()
 {
@@ -575,46 +505,40 @@ OptiqActions::ResetView()
     return true;
 }
 
-// The bookmark slots that currently hold a view.
 std::vector<int>
 OptiqActions::ListBookmarks() const
 {
     return m_trace_view != nullptr ? m_trace_view->ListBookmarks() : std::vector<int>();
 }
 
-// Stores the current zoom and scroll position in a numbered slot.
 bool
 OptiqActions::SaveBookmark(int slot)
 {
     return m_trace_view != nullptr && m_trace_view->SaveBookmark(slot);
 }
 
-// Restores the view saved in a slot.
 bool
 OptiqActions::GotoBookmark(int slot)
 {
     return m_trace_view != nullptr && m_trace_view->GotoBookmark(slot);
 }
 
-// Empties a bookmark slot.
 bool
 OptiqActions::RemoveBookmark(int slot)
 {
     return m_trace_view != nullptr && m_trace_view->RemoveBookmark(slot);
 }
 
-// Drops the two measurement pins on a span, the same as the Measure tool.
 bool
 OptiqActions::MeasureRange(double start_ns, double end_ns)
 {
-    if(m_trace_view == nullptr || !IsUsableRange(start_ns, end_ns))
+    if(m_trace_view == nullptr || !is_usable_time_range(start_ns, end_ns))
     {
         return false;
     }
     return m_trace_view->MeasureRange(start_ns, end_ns);
 }
 
-// Takes the measurement pins back off the timeline.
 bool
 OptiqActions::ClearMeasurement()
 {
@@ -626,11 +550,10 @@ OptiqActions::ClearMeasurement()
     return true;
 }
 
-// Zooms the visible window to a range, rather than only selecting it.
 bool
 OptiqActions::ZoomToRange(double start_ns, double end_ns)
 {
-    if(m_trace_view == nullptr || !IsUsableRange(start_ns, end_ns))
+    if(m_trace_view == nullptr || !is_usable_time_range(start_ns, end_ns))
     {
         return false;
     }
@@ -638,20 +561,18 @@ OptiqActions::ZoomToRange(double start_ns, double end_ns)
     return true;
 }
 
-// Pins a sticky note, which is saved with the project and outlives the chat.
 bool
 OptiqActions::AddNote(double time_ns, const std::string& title, const std::string& text,
                       double v_min, double v_max, uint64_t track_id)
 {
     if(m_trace_view == nullptr || !IsUsableInstant(time_ns) ||
-       !IsUsableRange(v_min, v_max))
+       !is_usable_time_range(v_min, v_max))
     {
         return false;
     }
     return m_trace_view->AddNote(time_ns, title, text, v_min, v_max, track_id);
 }
 
-// Raises a toast, for telling the user something without them reading the panel.
 void
 OptiqActions::Notify(const std::string& message, bool is_warning)
 {
@@ -659,11 +580,10 @@ OptiqActions::Notify(const std::string& message, bool is_warning)
         message, is_warning ? NotificationLevel::Warning : NotificationLevel::Info);
 }
 
-// Drag-selects a time range, the same as dragging on the ruler.
 bool
 OptiqActions::SelectRange(double start_ns, double end_ns)
 {
-    if(!HasTimeline() || !IsUsableRange(start_ns, end_ns))
+    if(!HasTimeline() || !is_usable_time_range(start_ns, end_ns))
     {
         return false;
     }
@@ -671,7 +591,6 @@ OptiqActions::SelectRange(double start_ns, double end_ns)
     return true;
 }
 
-// Drops the range selection.
 bool
 OptiqActions::ClearRange()
 {
@@ -683,13 +602,12 @@ OptiqActions::ClearRange()
     return true;
 }
 
-// Moves the visible window without changing the selection.
 bool
 OptiqActions::ShowRange(double start_ns, double end_ns)
 {
     // The timeline is what listens for this event, so require it rather than
     // just a data provider: a compute trace has the latter but not the former.
-    if(!HasTimeline() || !IsUsableRange(start_ns, end_ns))
+    if(!HasTimeline() || !is_usable_time_range(start_ns, end_ns))
     {
         return false;
     }
@@ -698,7 +616,6 @@ OptiqActions::ShowRange(double start_ns, double end_ns)
     return true;
 }
 
-// Scrolls a track into view on the timeline.
 bool
 OptiqActions::ScrollToTrack(uint64_t track_id)
 {
@@ -712,7 +629,6 @@ OptiqActions::ScrollToTrack(uint64_t track_id)
     return true;
 }
 
-// Expands and scrolls to a track in the topology sidebar.
 bool
 OptiqActions::RevealTrackInTopology(uint64_t track_id)
 {
@@ -725,8 +641,6 @@ OptiqActions::RevealTrackInTopology(uint64_t track_id)
     return true;
 }
 
-// The literal equivalent of clicking an event, which is what loads its details,
-// flow arrows, and call stack.
 bool
 OptiqActions::ClickEvent(uint64_t track_id, uint64_t event_uuid)
 {
@@ -739,7 +653,6 @@ OptiqActions::ClickEvent(uint64_t track_id, uint64_t event_uuid)
     return true;
 }
 
-// Adds an event to the selection, the same as clicking with multi-select held.
 bool
 OptiqActions::ShiftClickEvent(uint64_t track_id, uint64_t event_uuid)
 {
@@ -751,7 +664,6 @@ OptiqActions::ShiftClickEvent(uint64_t track_id, uint64_t event_uuid)
     return true;
 }
 
-// Deselects every event.
 bool
 OptiqActions::ClearEventSelection()
 {
@@ -763,7 +675,6 @@ OptiqActions::ClearEventSelection()
     return true;
 }
 
-// Makes an event glow until something clears it.
 bool
 OptiqActions::HighlightEvent(uint64_t track_id, uint64_t event_uuid)
 {
@@ -775,7 +686,6 @@ OptiqActions::HighlightEvent(uint64_t track_id, uint64_t event_uuid)
     return true;
 }
 
-// Turns off every persistent highlight.
 bool
 OptiqActions::ClearHighlights()
 {
@@ -787,7 +697,6 @@ OptiqActions::ClearHighlights()
     return true;
 }
 
-// Scrolls to an event and frames it, the same as jumping from a table row.
 bool
 OptiqActions::NavigateToEvent(uint64_t track_id, uint64_t event_uuid, double start_ns,
                               double duration_ns)
@@ -803,7 +712,6 @@ OptiqActions::NavigateToEvent(uint64_t track_id, uint64_t event_uuid, double sta
     return true;
 }
 
-// Selects a kernel in the compute views.
 bool
 OptiqActions::SelectKernel(uint32_t kernel_id)
 {
@@ -815,7 +723,6 @@ OptiqActions::SelectKernel(uint32_t kernel_id)
     return true;
 }
 
-// Selects a workload in the compute views.
 bool
 OptiqActions::SelectWorkload(uint32_t workload_id)
 {
