+++ src/view/src/agenticprofiling/rocprofvis_ai_tools_internal.h
@@ -8,20 +8,12 @@
  * surface is rocprofvis_ai_tools.h; nothing outside the agenticprofiling
  * directory should include this header.
  *
- * The tool bodies are split by what they touch. The ones in
- * rocprofvis_ai_ui_tools.cpp drive Optiq through OptiqActions and answer in the
- * same call; the ones in rocprofvis_ai_data_tools.cpp read the data model and
- * usually park a fetch for the panel to poll. That is the line worth drawing,
- * because it is also the line between a tool that can run on any trace and one
- * that has to reason about request ids and pending rows.
- * rocprofvis_ai_script_tools.cpp is a third case: the model writes Python that
- * computes the answer, so what comes back is a conclusion rather than rows.
- *
- * Each body file owns its own handler table and hands it to the dispatcher in
- * rocprofvis_ai_tools.cpp, so a tool body stays next to the entry that
- * registers it and the handlers themselves keep internal linkage. The handful
- * of helpers both halves need are declared here and defined beside the
- * dispatcher.
+ * Tool bodies are split by what they touch: rocprofvis_ai_ui_tools.cpp drives
+ * Optiq through OptiqActions and answers in the same call,
+ * rocprofvis_ai_data_tools.cpp reads the data model and usually parks a fetch
+ * for the panel to poll, and rocprofvis_ai_script_tools.cpp has the model write
+ * Python that computes the answer. Each body file owns its handler table and
+ * hands it to the dispatcher, so the handlers keep internal linkage.
  */
 
 #include "rocprofvis_ai_tools.h"
@@ -44,13 +36,11 @@ struct WorkloadInfo;
 // One tool body. Each takes the context for the trace in front and the
 // already-parsed arguments; the name it was called by is passed too, but only
 // the shared track_events / track_samples body has any use for it.
-using AssistantToolHandler = AssistantToolStartResult (*)(
+typedef AssistantToolStartResult (*AssistantToolHandler)(
     const AssistantToolContext& context, const jt::Json& args,
     const std::string& tool_name);
 
-// Maps the name the model used to the body that implements it. Kept beside the
-// bodies so adding a tool is one entry and one function, and so a name that is
-// registered in the schema but missing here fails loudly.
+// Maps the name the model used to the body that implements it.
 struct AssistantToolEntry
 {
     const char*          name;
@@ -65,17 +55,15 @@ struct AssistantToolTable
 };
 
 // The tables StartAssistantTool searches. Defined by the body files, so none of
-// them has to know the others exist.
+// them has to know the others exist. The script table is empty when scripting
+// is not built in, which is how that tool disappears without the dispatcher
+// knowing anything about the option.
 AssistantToolTable GetAssistantUiToolHandlers();
 AssistantToolTable GetAssistantDataToolHandlers();
-// Empty when scripting is not built in, which is how the tool disappears
-// without the dispatcher knowing anything about the option.
 AssistantToolTable GetAssistantScriptToolHandlers();
 
-// Formats a finished script run, or the decision that stopped it. Lives beside
-// the tool that started it, and is reached through FinishAssistantFetch like
-// every other fetch kind. AssistantScriptFetchPending is the panel-facing half
-// and is declared in rocprofvis_ai_tools.h.
+// Formats a finished script run, or the decision that stopped it. Reached
+// through FinishAssistantFetch like every other fetch kind.
 std::string FinishAssistantScriptFetch(const AssistantToolContext& context);
 
 // --- Helpers both halves need, defined beside the dispatcher ---------------
@@ -88,22 +76,30 @@ AssistantToolStartResult DoneResult(const std::string& content,
 uint64_t JsonU64(const jt::Json& json, const std::string& key,
                  uint64_t default_value);
 
+// Narrows a JSON number to an id, rejecting anything a uint64_t cannot hold.
+uint64_t JsonU64FromDouble(double value, uint64_t default_value);
+
+// Caps a tool result at max_chars, appending truncation_note when it had to
+// cut. Keeps the head, which is where a well-behaved result puts its summary.
+std::string TrimAssistantText(std::string text, size_t max_chars,
+                              const char* truncation_note);
+
 // Rejects a model-supplied array that is longer than anything a tool can use.
 // Returns true when the array is fine to walk.
 bool CheckArrayLength(const std::vector<jt::Json>& entries,
                       const char* argument_name, std::string& error_out);
 
 // Yields the selected time range, or the whole trace if nothing is selected.
-// A UI tool needs this as much as a query does: annotate stores the window the
-// user was looking at alongside the note.
 void SelectedOrFullTimeRange(const AssistantToolContext& context, double& start_ns,
                              double& end_ns);
 
 // Finds a kernel by id, then exact name, then a case-insensitive substring.
-// Shared because goto selects a kernel by the same loose name the model used to
-// ask kernel_metrics about it.
 const KernelInfo* FindComputeKernel(const WorkloadInfo& workload,
                                     const std::string& name, uint32_t id);
+
+// The workload a compute tool should read: the selected one, or the first
+// loaded when nothing is. Null when the trace has none at all.
+const WorkloadInfo* SelectedComputeWorkload(const AssistantToolContext& context);
 
 }  // namespace View
 }  // namespace RocProfVis
