+++ src/view/src/compute/rocprofvis_compute_view.cpp
@@ -102,9 +102,8 @@ ComputeView::ComputeView()
                 std::make_shared<TableDataEvent>(trace_path, request_id, response_code));
         });
 
-    // Same forwarding TraceView installs. Without it a compute tab runs
-    // requests that report no progress at all, which the script editor shows
-    // as a run stuck at zero.
+    // The forwarding TraceView installs. Without it a compute tab runs requests
+    // that report no progress at all.
     m_data_provider.SetRequestProgressUpdateCallback(
         [this](const RequestInfo& request, uint64_t pct, const std::string& message) {
             EventManager::GetInstance()->AddEvent(
@@ -114,7 +113,15 @@ ComputeView::ComputeView()
         });
 }
 
-ComputeView::~ComputeView() {}
+ComputeView::~ComputeView()
+{
+    // Every callback above captures this, and the provider outlives the view
+    // while its detached cleanup runs.
+    m_data_provider.SetTraceLoadedCallback(nullptr);
+    m_data_provider.SetFetchMetricsCallback(nullptr);
+    m_data_provider.SetTableDataReadyCallback(nullptr);
+    m_data_provider.SetRequestProgressUpdateCallback(nullptr);
+}
 
 std::optional<DataProviderCleanupWork>
 ComputeView::DetachProviderCleanup()
