+++ src/view/src/rocprofvis_data_provider.cpp
@@ -205,18 +205,13 @@ DataProvider::CleanupDetachedResources(DataProviderCleanupWork cleanup_work)
         if(req.request_obj_handle)
         {
 #ifdef ROCPROFVIS_ENABLE_SCRIPTING
+            // Cleanup runs after the tab and its editor are already gone, so
+            // ProcessExecuteScriptRequest never fires to free this.
             if(req.request_type == RequestType::kExecuteScript)
             {
                 rocprofvis_script_result_free(
                     reinterpret_cast<rocprofvis_controller_script_result_t*>(
                         req.request_obj_handle));
-                // Cleanup runs off the UI thread after the tab is already
-                // gone, so ProcessExecuteScriptRequest never fires. Queue
-                // the same event so the editor drops out of Running.
-                EventManager::GetInstance()->AddEvent(
-                    std::make_shared<ScriptExecuteCompleteEvent>(
-                        false, std::string(), "Trace closed",
-                        cleanup_result.trace_file_path));
             }
 #endif
             req.request_obj_handle = nullptr;
@@ -1341,155 +1336,6 @@ DataProvider::SaveTrimmedTrace(const std::string& path, double start_ns, double 
 
     return false;
 }
-
-#ifdef ROCPROFVIS_ENABLE_SCRIPTING
-rocprofvis_controller_arguments_t*
-DataProvider::BuildScriptContext(const std::vector<uint64_t>& track_ids, double start_ts,
-                                 double end_ts)
-{
-    rocprofvis_controller_arguments_t* context = rocprofvis_controller_arguments_alloc();
-    if(!context)
-    {
-        return nullptr;
-    }
-    rocprofvis_result_t result = rocprofvis_controller_set_double(
-        context, kRPVControllerScriptContextTimeRangeStart, 0, start_ts);
-    if(result == kRocProfVisResultSuccess)
-    {
-        result = rocprofvis_controller_set_double(
-            context, kRPVControllerScriptContextTimeRangeEnd, 0, end_ts);
-    }
-    uint64_t written = 0;
-    for(uint64_t track_id : track_ids)
-    {
-        rocprofvis_handle_t* track        = nullptr;
-        rocprofvis_result_t  track_result = rocprofvis_controller_get_object(
-            m_trace_controller, kRPVControllerSystemTrackById, track_id, &track);
-        if(track_result != kRocProfVisResultSuccess || !track)
-        {
-            continue;
-        }
-        track_result = rocprofvis_controller_set_object(
-            context, kRPVControllerScriptContextTracksIndexed, written, track);
-        if(track_result == kRocProfVisResultSuccess)
-        {
-            written++;
-        }
-    }
-    if(result == kRocProfVisResultSuccess)
-    {
-        result = rocprofvis_controller_set_uint64(
-            context, kRPVControllerScriptContextNumTracks, 0, written);
-    }
-    if(result != kRocProfVisResultSuccess)
-    {
-        rocprofvis_controller_arguments_free(context);
-        context = nullptr;
-    }
-    return context;
-}
-
-bool
-DataProvider::ExecuteScript(const std::string& source, const std::vector<uint64_t>& track_ids,
-                            double start_ts, double end_ts)
-{
-    if(m_state != ProviderState::kReady || !m_trace_controller)
-    {
-        spdlog::debug("Cannot execute script, provider not ready");
-        return false;
-    }
-    if(m_requests.find(EXECUTE_SCRIPT_REQUEST_ID) != m_requests.end())
-    {
-        spdlog::debug("Script already running");
-        return false;
-    }
-
-    // The previous run's output is not this run's output, and a caller that
-    // polls could otherwise read the old text as though the new script had
-    // already answered.
-    m_script_result_text.clear();
-    m_script_result_error.clear();
-    m_script_result_ok = false;
-
-    rocprofvis_controller_arguments_t* context =
-        BuildScriptContext(track_ids, start_ts, end_ts);
-    rocprofvis_controller_future_t*        future        = rocprofvis_controller_future_alloc();
-    rocprofvis_controller_script_result_t* script_result = nullptr;
-    if(!future)
-    {
-        rocprofvis_controller_arguments_free(context);
-        return false;
-    }
-
-    rocprofvis_result_t result = rocprofvis_script_execute_async(
-        m_trace_controller, source.c_str(), context, future, &script_result);
-    if(result != kRocProfVisResultSuccess || !script_result)
-    {
-        spdlog::error("script_execute_async failed ({})", static_cast<int>(result));
-        rocprofvis_controller_future_free(future);
-        rocprofvis_controller_arguments_free(context);
-        if(script_result)
-        {
-            rocprofvis_script_result_free(script_result);
-        }
-        return false;
-    }
-
-    RequestInfo request_info;
-    request_info.request_array      = nullptr;
-    request_info.request_future     = future;
-    request_info.request_obj_handle = script_result;
-    request_info.request_args       = context;
-    request_info.request_id         = EXECUTE_SCRIPT_REQUEST_ID;
-    request_info.loading_state      = RequestState::kLoading;
-    request_info.request_type       = RequestType::kExecuteScript;
-    request_info.request_progress   = 0;
-    m_requests.emplace(request_info.request_id, request_info);
-    return true;
-}
-
-bool
-DataProvider::CancelScript()
-{
-    return CancelRequest(EXECUTE_SCRIPT_REQUEST_ID);
-}
-
-void
-DataProvider::ProcessExecuteScriptRequest(RequestInfo& req)
-{
-    std::string text;
-    std::string error;
-    bool        success = req.response_code == kRocProfVisResultSuccess;
-    if(req.request_obj_handle)
-    {
-        text = GetString(req.request_obj_handle, kRPVControllerScriptResultText, 0);
-        error =
-            GetString(req.request_obj_handle, kRPVControllerScriptResultErrorMessage, 0);
-        rocprofvis_script_result_free(
-            reinterpret_cast<rocprofvis_controller_script_result_t*>(
-                req.request_obj_handle));
-        req.request_obj_handle = nullptr;
-    }
-    if(req.request_args)
-    {
-        rocprofvis_controller_arguments_free(req.request_args);
-        req.request_args = nullptr;
-    }
-    m_script_result_text  = text;
-    m_script_result_error = error;
-    m_script_result_ok    = success;
-    EventManager::GetInstance()->AddEvent(std::make_shared<ScriptExecuteCompleteEvent>(
-        success, text, error, GetTraceFilePath()));
-}
-
-bool
-DataProvider::GetLastScriptResult(std::string& text_out, std::string& error_out) const
-{
-    text_out  = m_script_result_text;
-    error_out = m_script_result_error;
-    return m_script_result_ok;
-}
-#endif
 
 bool
 DataProvider::CleanupDatabase(bool rebuild)
