+++ src/view/src/rocprofvis_data_provider.h
@@ -241,11 +241,10 @@ public:
                        double start_ts, double end_ts);
     bool CancelScript();
 
-    // What the last script on this provider produced, kept after its request
-    // is gone. The editor reads the completion event, but a caller that polls
-    // the request id instead - the assistant's script tool - has nothing left
-    // to read by the time it notices the request finished. Returns false when
-    // no script has run, or the script failed.
+    // What the last script produced, kept after its request is gone: a caller
+    // that polls the request id rather than listening for the completion event
+    // has nothing left to read by the time it notices the request finished.
+    // False when no script has run, or the script failed.
     bool GetLastScriptResult(std::string& text_out, std::string& error_out) const;
 #endif
 
@@ -377,8 +376,7 @@ private:
     // Current loading status progress in percents
     uint64_t m_progress_percent;
 #ifdef ROCPROFVIS_ENABLE_SCRIPTING
-    // Result of the last script, outliving its request. See
-    // GetLastScriptResult.
+    // Outlives the request it came from. See GetLastScriptResult.
     std::string m_script_result_text;
     std::string m_script_result_error;
     bool        m_script_result_ok = false;
