+++ src/view/src/rocprofvis_settings_panel.cpp
@@ -4,9 +4,6 @@
 #include "rocprofvis_settings_panel.h"
 #include "icons/rocprovfis_icon_defines.h"
 #include "imgui.h"
-#ifdef ROCPROFVIS_ENABLE_AGENTIC_PROFILING
-#    include "remote/rocprofvis_secret_store.h"
-#endif
 #include "rocprofvis_font_manager.h"
 #include "rocprofvis_settings_manager.h"
 #include "rocprofvis_utils.h"
@@ -170,10 +167,7 @@ SettingsPanel::Render()
                         ResetUnitOptions();
                         break;
                     }
-                    case Other:
-                    {
-                        break;
-                    }
+                    case Other: break;
 #ifdef ROCPROFVIS_ENABLE_AGENTIC_PROFILING
                     case Assistant:
                     {
@@ -421,200 +415,6 @@ SettingsPanel::RenderOtherSettings()
 
     m_settings_changed = true;
 }
-
-#ifdef ROCPROFVIS_ENABLE_AGENTIC_PROFILING
-// The Assistant page: endpoint URL, model, and the API key field.
-void
-SettingsPanel::RenderAssistantSettings()
-{
-    ImGuiStyle& style = ImGui::GetStyle();
-
-    AssistantSettings& assistant = m_usersettings.assistant;
-    if(assistant.providers.empty())
-    {
-        assistant.providers.push_back(MakeDefaultAssistantProvider());
-        assistant.active = 0;
-    }
-    if(assistant.active >= assistant.providers.size())
-    {
-        assistant.active = 0;
-    }
-    AssistantProvider& provider = assistant.providers[assistant.active];
-    ApplyAssistantEndpointDefaults(provider);
-
-    ImGui::TextUnformatted("Assistant");
-    ImGui::Separator();
-    ImGui::TextWrapped(
-        "OpenAI-compatible chat endpoint. The key is stored in the OS "
-        "credential store, never in the settings file.");
-
-    const float label_width =
-        ImGui::CalcTextSize("Model").x + 2.0f * style.ItemSpacing.x;
-
-    ImGui::AlignTextToFramePadding();
-    ImGui::TextUnformatted("URL");
-    ImGui::SameLine(label_width);
-    ImGui::SetNextItemWidth(-1.0f);
-    if(InputTextStringWithHint("##assistant_url", "https://<host>/<path>",
-                               provider.endpoint_url))
-    {
-        m_settings_changed = true;
-    }
-    if(ImGui::IsItemHovered())
-    {
-        SetTooltipStyled("Base URL, e.g. https://api.openai.com/v1. "
-                         "/chat/completions is appended. Azure-style /azure "
-                         "and /openai paths also insert the model as the "
-                         "deployment id.");
-    }
-
-    ImGui::AlignTextToFramePadding();
-    ImGui::TextUnformatted("Model");
-    ImGui::SameLine(label_width);
-    ImGui::SetNextItemWidth(-1.0f);
-    if(InputTextStringWithHint("##assistant_model", ASSISTANT_DEFAULT_MODEL,
-                               provider.model))
-    {
-        m_settings_changed = true;
-    }
-    if(ImGui::IsItemHovered())
-    {
-        SetTooltipStyled("Model name sent to the endpoint. On Azure-style URLs "
-                         "this is the deployment id. Default is %s.",
-                         ASSISTANT_DEFAULT_MODEL);
-    }
-
-    ImGui::Dummy(ImVec2(0.0f, style.ItemSpacing.y));
-    ImGui::TextUnformatted("API key");
-    ImGui::Separator();
-
-    const bool token_stored =
-        m_settings.HasAssistantToken(provider.name) && !m_assistant_clear_token;
-    if(token_stored && m_assistant_token_draft.empty())
-    {
-        ImGui::TextUnformatted("A key is stored. Leave the field blank to keep it.");
-    }
-    else if(SecretStore::IsAvailable())
-    {
-        ImGui::TextWrapped(
-            "Saved to the OS credential store when you press OK, never to the "
-            "settings file.");
-    }
-    else
-    {
-        ImGui::TextWrapped(
-            "No OS credential store is available. The key is kept in memory "
-            "for this session only.");
-    }
-
-    ImGui::AlignTextToFramePadding();
-    ImGui::TextUnformatted("Key");
-    ImGui::SameLine();
-    const float eye_w = ImGui::GetFrameHeight();
-    ImGui::SetNextItemWidth(-(eye_w + style.ItemInnerSpacing.x));
-    ImGuiInputTextFlags token_flags =
-        m_assistant_show_token ? 0 : ImGuiInputTextFlags_Password;
-    if(InputTextStringWithHint("##assistant_token",
-                               token_stored ? "Stored — type to replace" : "API key",
-                               m_assistant_token_draft, token_flags))
-    {
-        m_assistant_clear_token = false;
-        m_settings_changed      = true;
-    }
-    ImGui::SameLine(0.0f, style.ItemInnerSpacing.x);
-    ImFont* icon_font = m_fonts.GetFont(FontType::kIcon);
-    ImGui::PushID("assistant_token_reveal");
-    if(IconButton(m_assistant_show_token ? ICON_EYE_SLASH : ICON_EYE, icon_font,
-                  ImVec2(eye_w, eye_w), m_assistant_show_token ? "Hide" : "Show"))
-    {
-        m_assistant_show_token = !m_assistant_show_token;
-    }
-    ImGui::PopID();
-
-    if(ImGui::Button("Clear stored key"))
-    {
-        m_assistant_token_draft.clear();
-        m_assistant_clear_token = true;
-        m_settings_changed      = true;
-    }
-}
-
-// Commits the pending key edits. Called from OK, so nothing typed on this page
-// reaches the credential store until the user accepts the dialog.
-void
-SettingsPanel::ApplyAssistantTokenEdits()
-{
-    // Endpoints dropped by Reset go first: their keys have nothing pointing at
-    // them once the configuration is gone.
-    for(const std::string& orphan : m_assistant_orphaned_providers)
-    {
-        m_settings.ClearAssistantToken(orphan);
-    }
-
-    if(m_usersettings.assistant.active < m_usersettings.assistant.providers.size())
-    {
-        ApplyAssistantEndpointDefaults(
-            m_usersettings.assistant.providers[m_usersettings.assistant.active]);
-    }
-
-    const std::string provider_name = ActiveAssistantProviderName();
-    if(m_assistant_clear_token)
-    {
-        m_settings.ClearAssistantToken(provider_name);
-    }
-    else if(!m_assistant_token_draft.empty())
-    {
-        m_settings.SetAssistantToken(provider_name, m_assistant_token_draft);
-    }
-    DiscardAssistantTokenEdits();
-}
-
-// Throws away the pending key edits without touching the credential store.
-void
-SettingsPanel::DiscardAssistantTokenEdits()
-{
-    m_assistant_token_draft.clear();
-    m_assistant_orphaned_providers.clear();
-    m_assistant_clear_token = false;
-}
-
-// Name of the endpoint being edited, which is the key its token is stored under.
-std::string
-SettingsPanel::ActiveAssistantProviderName() const
-{
-    const AssistantSettings& assistant = m_usersettings.assistant;
-    if(assistant.active >= assistant.providers.size())
-    {
-        return std::string();
-    }
-    return assistant.providers[assistant.active].name;
-}
-
-// Restores the default endpoint and drops the pending key edit. Endpoints that
-// disappear here are remembered so OK can delete their keys too, rather than
-// leaving entries in the credential store that nothing refers to any more.
-void
-SettingsPanel::ResetAssistantOptions()
-{
-    const AssistantProvider default_provider = MakeDefaultAssistantProvider();
-    for(const AssistantProvider& provider : m_usersettings.assistant.providers)
-    {
-        if(provider.name != default_provider.name)
-        {
-            m_assistant_orphaned_providers.push_back(provider.name);
-        }
-    }
-
-    m_usersettings.assistant.providers.clear();
-    m_usersettings.assistant.providers.push_back(default_provider);
-    m_usersettings.assistant.active = 0;
-    m_assistant_token_draft.clear();
-    // The default endpoint is back to having nothing configured, so its key
-    // should go as well.
-    m_assistant_clear_token = true;
-    m_settings_changed      = true;
-}
-#endif  // ROCPROFVIS_ENABLE_AGENTIC_PROFILING
 
 void
 SettingsPanel::ResetDisplayOptions()
