+++ src/view/src/rocprofvis_trace_view.cpp
@@ -41,37 +41,6 @@ namespace View
 constexpr ImVec2 MINIMAP_POPUP_SIZE(600.0f, 435.0f);
 constexpr ImVec2 MINIMAP_POPUP_MIN_SIZE(500.0f, 360.0f);
 
-// Notes dropped by the assistant land near the top of the track area at a
-// readable default size; the user can drag and resize them from there.
-constexpr float  ANNOTATION_DEFAULT_Y_OFFSET = 40.0f;
-constexpr ImVec2 ANNOTATION_DEFAULT_SIZE(260.0f, 120.0f);
-
-// Bookmark slots exposed in the toolbar dropdown.
-constexpr int BOOKMARK_MIN_SLOT = 0;
-constexpr int BOOKMARK_MAX_SLOT = 9;
-
-namespace
-{
-
-// Slots outside the dropdown's range are rejected everywhere, not just on save,
-// so a bad slot fails the same way whichever entry point it arrives through.
-bool
-IsValidBookmarkSlot(int slot)
-{
-    return slot >= BOOKMARK_MIN_SLOT && slot <= BOOKMARK_MAX_SLOT;
-}
-
-// True when a time range can be acted on. Non-finite values have to be caught
-// explicitly: comparing anything against NaN is false, so "end <= start" alone
-// would let a NaN through into the view transform.
-bool
-IsUsableTimeRange(double start_ns, double end_ns)
-{
-    return std::isfinite(start_ns) && std::isfinite(end_ns) && end_ns > start_ns;
-}
-
-}  // namespace
-
 TraceView::TraceView()
 : m_timeline_view(nullptr)
 , m_horizontal_split_container(nullptr)
@@ -235,6 +204,7 @@ TraceView::~TraceView()
     m_data_provider.SetTraceLoadedCallback(nullptr);
     m_data_provider.SetSaveTraceCallback(nullptr);
     m_data_provider.SetCleanupDatabaseCallback(nullptr);
+    m_data_provider.SetRequestProgressUpdateCallback(nullptr);
 
     EventManager::GetInstance()->Unsubscribe(static_cast<int>(RocEvents::kTabSelected),
                                              m_tabselected_event_token);
@@ -732,226 +702,6 @@ TraceView::SetHistogramVisibility(bool visibility)
     }
 }
 
-void
-TraceView::SetMinimapVisibility(bool visibility)
-{
-    m_show_minimap_popup = visibility;
-}
-
-bool
-TraceView::IsMinimapVisible() const
-{
-    return m_show_minimap_popup;
-}
-
-void
-TraceView::SetFlowArrowsVisible(bool visible)
-{
-    if(m_timeline_view)
-    {
-        m_timeline_view->GetArrowLayer().SetFlowDisplayMode(
-            visible ? FlowDisplayMode::kShowAll : FlowDisplayMode::kHide);
-    }
-}
-
-bool
-TraceView::AreFlowArrowsVisible() const
-{
-    return m_timeline_view && m_timeline_view->GetArrowLayer().GetFlowDisplayMode() ==
-                                  FlowDisplayMode::kShowAll;
-}
-
-void
-TraceView::SetFlowRenderChained(bool chained)
-{
-    if(m_timeline_view)
-    {
-        m_timeline_view->GetArrowLayer().SetRenderStyle(
-            chained ? TimelineArrow::RenderStyle::kChain
-                    : TimelineArrow::RenderStyle::kFan);
-    }
-}
-
-std::vector<std::string>
-TraceView::ListAnalysisTabs()
-{
-    return m_analysis_view ? m_analysis_view->ListTabs() : std::vector<std::string>();
-}
-
-bool
-TraceView::SelectAnalysisTab(const std::string& name)
-{
-    if(!m_analysis_view || !m_analysis_view->SelectTab(name))
-    {
-        return false;
-    }
-    // A tab the user cannot see is not a switch, so make sure the panel is up.
-    SetAnalysisViewVisibility(true);
-    SettingsManager::GetInstance().GetAppWindowSettings().show_details_panel = true;
-    return true;
-}
-
-std::string
-TraceView::ActiveAnalysisTab()
-{
-    return m_analysis_view ? m_analysis_view->ActiveTab() : std::string();
-}
-
-#ifdef ROCPROFVIS_ENABLE_SCRIPTING
-bool
-TraceView::ProposeScript(const std::string& source)
-{
-    if(!m_analysis_view || !m_analysis_view->GetScriptEditor())
-    {
-        return false;
-    }
-    m_analysis_view->GetScriptEditor()->ProposeScript(source);
-    return true;
-}
-
-ScriptApproval
-TraceView::ScriptProposalState() const
-{
-    if(!m_analysis_view || !m_analysis_view->GetScriptEditor())
-    {
-        return ScriptApproval::kNone;
-    }
-    return m_analysis_view->GetScriptEditor()->ProposalState();
-}
-
-void
-TraceView::ClearScriptProposal()
-{
-    if(m_analysis_view && m_analysis_view->GetScriptEditor())
-    {
-        m_analysis_view->GetScriptEditor()->ClearProposal();
-    }
-}
-#endif
-
-void
-TraceView::ResetView()
-{
-    if(m_timeline_view)
-    {
-        const TimelineModel& timeline = m_data_provider.DataModel().GetTimeline();
-        m_timeline_view->MoveToPosition(timeline.GetStartTime(), timeline.GetEndTime(),
-                                        0.0, false);
-    }
-}
-
-void
-TraceView::SetAnnotationsVisible(bool visible)
-{
-    if(m_annotations)
-    {
-        m_annotations->SetVisible(visible);
-    }
-}
-
-bool
-TraceView::AreAnnotationsVisible() const
-{
-    return m_annotations && m_annotations->IsVisibile();
-}
-
-std::vector<int>
-TraceView::ListBookmarks() const
-{
-    std::vector<int> slots;
-    slots.reserve(m_bookmarks.size());
-    for(const std::pair<const int, ViewCoords>& bookmark : m_bookmarks)
-    {
-        slots.push_back(bookmark.first);
-    }
-    std::sort(slots.begin(), slots.end());
-    return slots;
-}
-
-bool
-TraceView::SaveBookmark(int slot)
-{
-    if(!m_timeline_view || !IsValidBookmarkSlot(slot))
-    {
-        return false;
-    }
-    m_bookmarks[slot] = m_timeline_view->GetViewCoords();
-    return true;
-}
-
-bool
-TraceView::GotoBookmark(int slot)
-{
-    if(!IsValidBookmarkSlot(slot))
-    {
-        return false;
-    }
-    const std::unordered_map<int, ViewCoords>::const_iterator it = m_bookmarks.find(slot);
-    if(it == m_bookmarks.end() || !m_timeline_view)
-    {
-        return false;
-    }
-    m_timeline_view->MoveToPosition(it->second.v_min_x, it->second.v_max_x, it->second.y,
-                                    false);
-    return true;
-}
-
-bool
-TraceView::RemoveBookmark(int slot)
-{
-    return IsValidBookmarkSlot(slot) && m_bookmarks.erase(slot) > 0;
-}
-
-bool
-TraceView::MeasureRange(double start_ns, double end_ns)
-{
-    if(!m_measurement || !IsUsableTimeRange(start_ns, end_ns))
-    {
-        return false;
-    }
-    // EnterMeasurementMode keeps any existing pins, so clear before dropping the
-    // new pair or the second point would land in the wrong slot.
-    m_measurement->EnterMeasurementMode();
-    m_measurement->ClearMeasurement();
-    m_measurement->SetFreehandMode(true);
-    m_measurement->SetFreehandMeasurementPoint(start_ns);
-    m_measurement->SetFreehandMeasurementPoint(end_ns);
-    return true;
-}
-
-void
-TraceView::ClearMeasurement()
-{
-    if(m_measurement)
-    {
-        m_measurement->ClearMeasurement();
-        m_measurement->ExitMeasurementMode();
-    }
-}
-
-void
-TraceView::ZoomToRange(double start_ns, double end_ns)
-{
-    if(m_timeline_view && IsUsableTimeRange(start_ns, end_ns))
-    {
-        m_timeline_view->SetViewableRangeNS(start_ns, end_ns);
-    }
-}
-
-bool
-TraceView::AddNote(double time_ns, const std::string& title, const std::string& text,
-                   double v_min, double v_max, uint64_t track_id)
-{
-    if(!m_annotations || !std::isfinite(time_ns) || !IsUsableTimeRange(v_min, v_max))
-    {
-        return false;
-    }
-    m_annotations->SetVisible(true);
-    m_annotations->AddSticky(time_ns, ANNOTATION_DEFAULT_Y_OFFSET,
-                             ANNOTATION_DEFAULT_SIZE, text, title, v_min, v_max,
-                             track_id, false, false);
-    return true;
-}
 
 void
 TraceView::RenderToolbar()
