+++ src/view/src/widgets/rocprofvis_script_editor.h
@@ -18,16 +18,14 @@ namespace View
 class DataProvider;
 class TimelineSelection;
 
-// Where a script the assistant offered has got to. A script it wrote is the
-// one thing it does that runs real code, so it never runs unattended: the
-// editor holds it until the user reads it and decides.
+// Where a script the assistant offered has got to. A script it wrote never
+// runs unattended: the editor holds it until the user reads it and decides.
 enum class ScriptApproval : uint8_t
 {
     // Nothing offered. Manual runs stay in this state throughout.
     kNone,
     kPending,
     kRejected,
-    // Approved and handed to the interpreter.
     kRunning,
     kFinished,
     // Approved, but the run could not be started at all.
@@ -38,10 +36,9 @@ enum class ScriptApproval : uint8_t
  * @brief The Script tab of the details panel: a Python editor over the trace
  * that owns it.
  *
- * One per trace, built and owned by AnalysisView like the other tabs, so the
- * source, the output, and any run belong to the trace the user is looking at.
- * That is also what lets Run go straight to its own DataProvider rather than
- * working out which tab is in front.
+ * One per trace, built and owned by AnalysisView like the other tabs, so Run
+ * goes straight to its own DataProvider rather than working out which tab is
+ * in front.
  */
 class ScriptEditor : public RocWidget
 {
@@ -57,41 +54,36 @@ public:
     // every UI change the assistant makes goes through one place.
     void ProposeScript(const std::string& source);
 
-    // Where that offer has got to, and how the caller lets go of it once it has
-    // read the answer.
     ScriptApproval ProposalState() const;
     void           ClearProposal();
 
 private:
-    void Run();
-    void Cancel();
-    void Reject();
-    void LoadFromFile();
-    void SaveToFile();
-    void ReadFile(const std::string& path);
-    void WriteFile(const std::string& path);
-    // Title, subtitle and the action row, in the header band the rest of the
-    // app's panels use.
-    void RenderHeaderCard();
-    // Only while a script is waiting on the user: says where it came from and
-    // what pressing Run would do.
-    void RenderProposalBanner();
-    void RenderSource(const ImVec2& size);
-    void RenderOutput(const ImVec2& size);
-    bool CanRun() const;
-    // What the subtitle under the title says: the loaded file, or how the
-    // script got here.
+    void        Run();
+    void        Cancel();
+    void        Reject();
+    void        LoadFromFile();
+    void        SaveToFile();
+    void        ReadFile(const std::string& path);
+    void        WriteFile(const std::string& path);
+    void        RenderHeaderCard();
+    void        RenderProposalBanner();
+    void        RenderSource(const ImVec2& size);
+    void        RenderOutput(const ImVec2& size);
+    // split_extent is the workspace along the axis being split; cross_length is
+    // the handle's length across it.
+    void        RenderSplitter(bool columns, float split_extent, float cross_length);
+    bool        CanRun() const;
     std::string SubtitleText() const;
 
     DataProvider&                      m_data_provider;
     std::shared_ptr<TimelineSelection> m_timeline_selection;
 
     bool        m_running;
+    // Share of the workspace given to the Result pane, as dragged.
+    float       m_result_ratio;
     uint64_t    m_progress_percent;
     std::string m_source;
     std::string m_output;
-    // Tints the output pane, so a traceback reads as a failure at a glance
-    // instead of looking like results.
     bool        m_output_is_error;
     std::string m_status;
     std::string m_file_path;
